refactor: ValidationUtils 패키지 변경
util 패키지 → domain 패키지 변경
domain 패키지에 속한 클래스의 데이터를 검증한다
