+++ src/main/java/domain/Movement.java
@@ -1,7 +1,5 @@
 package domain;
 
-import utils.ValidationUtils;
-
 public class Movement {
     private int position;
