+++ src/test/java/domain/ValidationUtilsTest.java
@@ -1,5 +1,6 @@
-package utils;
+package domain;
 
+import domain.ValidationUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
